+++ NEWS.md
@@ -1,3 +1,14 @@
+## CalibrateSSB 1.4.0
+* New pkgdown website for the package  
+  - This package now has a documentation site at [https://statisticsnorway.github.io/ssb-calibratessb/](https://statisticsnorway.github.io/ssb-calibratessb/).
+* New parameters to `CalibrateSSB()`: `ginvtol`, `boundstol` and `resids_by_lm`
+  - Control of robustness and precision in regression computations.
+* Support for tibble and data.table input to `CalibrateSSB()` (`grossSample`, `popData` and `popTotals`)
+  - Internal coercion to data.frame using `as.data.frame()` where necessary.
+* Bugfix: `samplingWeightsOutput = TRUE` in `CalibrateSSB()` now works correctly
+  - Correct output of sampling-weight data instead of the `samplingWeights` variable name.
+  - Avoidance of downstream errors in `WideFromCalibrate()` and `CalibrateSSBpanel()`.
+
 
 ## CalibrateSSB	1.3.0